pyproject.toml
[project]
name = "webmaster-helper"
version = "0.0.1"
description = "This is the simple CRM system for webmasters to collect all business contacts and collect site SEO metrics."
authors = [
    {name = "alexmudrak", email = "tiit.learn@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
    "psycopg2-binary>=2.9.7",
    "python-dotenv>=1.0.0",
    "django-filter>=23.3",
    "djangorestframework-simplejwt>=5.3.0",
    "setuptools>=68.2.2",
    "django-cors-headers>=4.2.0",
    "celery>=5.3.4",
    "Redis>=5.0.1",
    "playwright>=1.38.0",
    "playwright-stealth>=1.0.6",
    "httpx>=0.25.0",
    "fake-useragent>=1.3.0",
    "lxml>=4.9.3",
    "gunicorn>=21.2.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
_.env_file = ".env"
init = "pre-commit install"
dev = "python src/manage.py runserver"
prod = "gunicorn --bind 0.0.0.0:8000 src.core.wsgi:application"
manager = "python src/manage.py"
coverage = "coverage run --branch -m pytest tests"
coverage-report = "coverage report --include=src/** --show-missing"
coverage-report-html = "coverage html --include=src/**"
coverage-report-ci = "coverage report --include=src/** --show-missing --fail-under=80"

[tool.pdm.dev-dependencies]
dev = [
    "coverage>=7.3.1",
    "pytest>=7.4.2",
    "pytest-asyncio>=0.21.1",
    "pytest-django>=4.5.2",
]

[tool.black]
line-length = 79

[tool.isort]
line_length = 79
profile = "black"

[tool.ruff]
line-length = 79
fixable = ["ALL"]
unfixable = []

[tool.coverage.report]
exclude_lines= ["pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "raise Exception",
    "if __name__ == .__main__.:"]

omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/ui/*",
    "*/migrations/*",
    "*/src/core/*",
    "*/src/manage.py",
    ]

[tool.coverage.run]
source = ["src"]

omit = [
    "*/.virtualenvs/*",
    "*/virtualenvs/*",
    "*/tests/*",
    "*/.local/*",
    "*/__init__.py",
    "*/ui/*",
    "*/migrations/*",
    "*/src/core/*",
    "*/src/manage.py",
    ]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
norecursedirs = ["database"]
